Style the Edit Page
